requirements.txt
beautifulsoup4==4.12.3
celery[redis]
Django==5.1.1
djangorestframework==3.15.2
requests==2.32.2
